requirements.txt
quart
hypercorn
tortoise-orm
pydantic[dotenv]
